Added buttons on sounds explorer window to mute all attached / rezzed sounds by avatar

+++ indra/newview/fsassetblacklist.cpp
@@ -72,7 +72,7 @@ void FSAssetBlacklist::init()
     loadBlacklist();
 }
 
-bool FSAssetBlacklist::isBlacklisted(const LLUUID& id, LLAssetType::EType type)
+bool FSAssetBlacklist::isBlacklisted(const LLUUID& id, LLAssetType::EType type, eBlacklistFlag flag)
 {
     if (mBlacklistData.empty())
     {
@@ -87,29 +87,71 @@ bool FSAssetBlacklist::isBlacklisted(const LLUUID& id, LLAssetType::EType type)
     }
 
     blacklisted_uuid_container_t uuids = it->second;
-    return (uuids.find(id) != uuids.end());
-}
-
-void FSAssetBlacklist::addNewItemToBlacklist(const LLUUID& id, const std::string& name, const std::string& region, LLAssetType::EType type, bool permanent /*= true*/, bool save /*= true*/)
-{
-    if (isBlacklisted(id, type))
+    if (uuids.find(id) == uuids.end())
     {
-        return;
+        return false;
     }
 
+    if (flag == eBlacklistFlag::NONE)
+    {
+        return true;
+    }
+
+    const auto& data_it = mBlacklistData.find(id);
+    if (data_it == mBlacklistData.end())
+    {
+        return false;
+    }
+
+    const LLSD& data = data_it->second;
+    if (!data.has("asset_blacklist_flag"))
+    {
+        return false;
+    }
+
+    eBlacklistFlag stored_flag = static_cast<eBlacklistFlag>(data["asset_blacklist_flag"].asInteger());
+
+    return (static_cast<int>(stored_flag) & static_cast<int>(flag)) != 0;
+}
+
+void FSAssetBlacklist::addNewItemToBlacklist(const LLUUID& id, const std::string& name, const std::string& region, LLAssetType::EType type, eBlacklistFlag flag /*= eBlacklistFlag::NONE*/, bool permanent /*= true*/, bool save /*= true*/)
+{
     LLDate curdate = LLDate((double)time_corrected());
     std::string input_date = curdate.asString();
     input_date.replace(input_date.find("T"), 1, " ");
     input_date.resize(input_date.size() - 1);
 
     LLSD data;
-    data["asset_name"] = name;
-    data["asset_region"] = region;
-    data["asset_type"] = type;
-    data["asset_date"] = input_date;
-    data["asset_permanent"] = permanent;
 
-    addNewItemToBlacklistData(id, data, save);
+    if (isBlacklisted(id, type))
+    {
+        auto it = mBlacklistData.find(id);
+        if (it != mBlacklistData.end())
+        {
+            data = it->second;
+
+            S32 existing_flag = data.has("asset_blacklist_flag") ? data["asset_blacklist_flag"].asInteger() : 0;
+            data["asset_blacklist_flag"] = static_cast<S32>(existing_flag | static_cast<S32>(flag));
+
+            data["asset_name"] = name;
+            data["asset_region"] = region;
+            data["asset_date"] = input_date;
+            data["asset_permanent"] = permanent;
+
+            addNewItemToBlacklistData(id, data, save);
+        }
+    }
+    else
+    {
+        data["asset_name"] = name;
+        data["asset_region"] = region;
+        data["asset_type"] = type;
+        data["asset_blacklist_flag"] = static_cast<S32>(flag);
+        data["asset_date"] = input_date;
+        data["asset_permanent"] = permanent;
+
+        addNewItemToBlacklistData(id, data, save);
+    }
 }
 
 bool FSAssetBlacklist::removeItem(const LLUUID& id)
@@ -122,11 +164,14 @@ bool FSAssetBlacklist::removeItem(const LLUUID& id)
     {
         return false;
     }
+    
+    // Erase for each possible type
+    for (auto& [type, container] : mBlacklistTypeContainer)
+    {
+        container.erase(id);
+    }
 
     LLSD data = it->second;
-    LLAssetType::EType type = S32toAssetType(data["asset_type"].asInteger());
-
-    mBlacklistTypeContainer[type].erase(id);
     mBlacklistData.erase(it);
 
     return data["asset_permanent"].asBoolean();
@@ -165,14 +210,50 @@ void FSAssetBlacklist::removeItemsFromBlacklist(const uuid_vec_t& ids)
     }
 }
 
+void FSAssetBlacklist::removeFlagsFromItem(const LLUUID& id, S32 combined_flags)
+{auto it = mBlacklistData.find(id);
+    if (it == mBlacklistData.end())
+    {
+        return;
+    }
+
+    LLSD& data = it->second;
+    S32 current_flags = data.has("asset_blacklist_flag") ? data["asset_blacklist_flag"].asInteger() : 0;
+
+    current_flags &= ~combined_flags;
+
+    if (current_flags == 0)
+    {
+        removeItemsFromBlacklist({ id });
+    }
+    else
+    {
+        data["asset_blacklist_flag"] = current_flags;
+        addNewItemToBlacklistData(id, data, true);
+
+        if (!mBlacklistChangedCallback.empty())
+        {
+            mBlacklistChangedCallback(LLSD().with(id.asString(), data), eBlacklistOperation::BLACKLIST_ADD);
+        }
+    }
+}
+
 void FSAssetBlacklist::addNewItemToBlacklistData(const LLUUID& id, const LLSD& data, bool save)
 {
     LLAssetType::EType type = S32toAssetType(data["asset_type"].asInteger());
 
-    addEntryToBlacklistMap(id, type);
-    mBlacklistData[id] = data;
+    auto it = mBlacklistData.find(id);
+    if (it != mBlacklistData.end())
+    {
+        it->second = data;
+    }
+    else
+    {
+        addEntryToBlacklistMap(id, type);
+        mBlacklistData[id] = data;
+    }
 
-    if (type == LLAssetType::AT_SOUND)
+    if (type == LLAssetType::AT_SOUND && data["asset_blacklist_flag"].asInteger() == 0)
     {
         LLFileSystem::removeFile(id, LLAssetType::AT_SOUND);
         std::string wav_path = gDirUtilp->getExpandedFilename(LL_PATH_FS_SOUND_CACHE, id.asString()) + ".dsf";
